Fixing an occasional error in commands.
Some commands were returning an error akin to:
Command 'hunt' raised an exception: TypeError: ChannelManagement._error_if_not_bot_channel() missing 1 required keyword-only argument: 'message'

Making the message argument truly optional.
